feat: update user update service

+++ backend/src/controllers/user.js
@@ -69,13 +69,25 @@ class UserController {
     try {
       const user_id = req.params.id;
 
-      const user = userRepository.findById(user_id);
+      const user = await userRepository.findById(user_id);
 
-      const updated_user = userRepository.update(user);
+      if (!user) {
+        return responseHandler.notFoundResponse(
+          res,
+          "User with id: " + user_id + " not found"
+        );
+      }
+
+      const input = req.body;
+
+      const user_input = new User(user_id, input.name ?? user.name, input.email ?? user.email, input.password ?? user.password);
+
+      await userRepository.update(user_input);
+
       return responseHandler.successResponseWithData(
         res,
         "Update User with id: " + user_id,
-        updated_user,
+        user_input,
         200
       );
     } catch (error) {
@@ -88,6 +100,13 @@ class UserController {
       const user_id = req.params.id;
 
       const user = await userRepository.findById(user_id);
+
+      if (!user) {
+        return responseHandler.notFoundResponse(
+          res,
+          "User with id: " + user_id + " not found"
+        );
+      }
 
       await userRepository.delete(user_id);
